feat: create the TextTypes to hold common types about text component that will be re-used
