+++ react-client/src/components/design-system/text/Text.tsx
@@ -1,28 +1,6 @@
-import { ReactNode } from "react";
-import { DisplayClasses } from "~/routes/types/styleTypes";
-import {
-  TextCommonStyleProps,
-  useTextCommonStyles,
-} from "./useTextCommonStyles";
+import { useTextCommonStyles } from "./useTextCommonStyles";
 import clsx from "clsx";
-
-type StyleProps = TextCommonStyleProps & {
-  display?: DisplayClasses;
-  inherit?: boolean;
-  fontFamily?: "franklin" | "raleway" | "roboto";
-  fontWeight?: "normal" | "bold" | "semibold" | "medium";
-  isResponsive?: boolean;
-  size?: "12" | "14" | "16" | "18" | "20" | "24" | "32" | "40";
-  textDecoration?: "none" | "line-through" | "underline";
-};
-
-export type TextProps = StyleProps & {
-  ariaHidden?: boolean;
-  ariaLabel?: string;
-  children: ReactNode;
-  element?: "p" | "span";
-  id?: string;
-};
+import { StyleProps, TextProps } from "../types/TextTypes";
 
 export const Text = ({
   ariaHidden,
